여러가지 예제 추가
