+++ examples/counter/index.html
