+++ examples/counter/style.css
@@ -10,7 +10,13 @@ body {
 .counter-wrapper {
   display: flex;
   flex-direction: column;
+  align-items: center;
   width: 100%;
   height: 100%;
   max-width: 480px;
+}
+
+img {
+  width: 150px;
+  height: 150px;
 }
