+++ package.json
@@ -4,8 +4,10 @@
   "description": "Yet Another Simple JSX using tagged template",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "parcel src/index.html",
-    "build": "parcel build src/index.html"
+    "counter": "parcel examples/counter/index.html",
+    "recursion": "parcel examples/recursion/index.html",
+    "input": "parcel examples/input/index.html",
+    "build": "parcel build examples/counter/index.html"
   },
   "alias": {
     "src": "./src"
